Update - v1.0.0

+++ src/utils/progressbar.h
@@ -1,6 +1,7 @@
 #ifndef _utils_progressbar_h
 #define _utils_progressbar_h
 
+#include "utils.h"
 #include <stdint.h>
 #include <string>
 
@@ -20,7 +21,7 @@ public:
     {
         std::string output = this->barPrefix;
         output += "[";
-        for (int i = 0; i < 50; i++)
+        for (size_t i = 0; i < 50; i++)
             output += (i < (progress / 2) ? "■" : "-");
         output += "]";
         return string_format("%s %s", output.c_str(), sufix.c_str());
